WIP: try to implmenet work stealin DFS
I tried.. but it seems that sharing complex data type
among the processes is not simple with MPI. I stopped
when implmeneting WorkStealingStack. Becauuse I would like
to have a buffer of complex data types (Board), but
I cant. The other possibility is to share only puzzle states, thus
int[81], but that would carry a overhead of initzializing the
Board instance or reimplement the whole algorithm. That would probably
also result in performance decrease..